Jekyll 2.0 Rebuild
This is a major re-build of the site (and the launch of a new branch)
based on Jekyll 2.0's default blog.

A number of important changes have been made:
- Gridism framework stripped out (responsive grid elements will be
constructed where needed)
- default responsive nag is big improvement over previous versions
- removal of discreet post sub-folders – Jekyll categories and the new
collections features should accomplish all that is needed on this front

Prose and GH-Pages integration will need to be reconfigured (config.yml
is mostly based on the Jekyll defaults again).